feat(output-view, racing-game): 승자는 구하는 로직을 outputview 에서 domain 이 담당하도록 변경
before:
승자를 구하는 로직과 print 기능을 모두 outputview에서 담당하고 있음
after:
승자를 구하는 로직은 domain, 결과를 출력하는 로직은 outputview에서 담당
