+++ src/main/java/domain/RacingGame.java
@@ -1,11 +1,16 @@
 package domain;
 
 import car.Car;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import view.OutputView;
 
 public class RacingGame {
 
+    private static final String DELIMITER = ",";
     private final List<Car> carNames;
     private final int round;
     private OutputView message = new OutputView();
@@ -26,10 +31,25 @@ public class RacingGame {
     }
 
     public void getWinner() {
-        printWinner(carNames);
+        String winner = findWinners(carNames);
+        message.printWinner(winner);
     }
 
-    private void printWinner(List<Car> carNames) {
-        message.printWinners(carNames);
+    private String findWinners(List<Car> cars) {
+        final Map<Integer, List<String>> scoreBoard;
+
+        scoreBoard = makeScoreBoard(cars);
+        final int maxPosition = Collections.max(scoreBoard.keySet());
+        return String.join(DELIMITER, scoreBoard.get(maxPosition));
+    }
+
+    private Map<Integer, List<String>> makeScoreBoard(List<Car> cars) {
+        final Map<Integer, List<String>> scoreBoard = new HashMap<>();
+        for (Car racingCar : cars) {
+            final List<String> names =
+                scoreBoard.computeIfAbsent(racingCar.getPosition(), ArrayList::new);
+            names.add(racingCar.getName());
+        }
+        return scoreBoard;
     }
 }
